requirements.txt
alembic==0.8.3
blinker==1.4
decorator==4.0.4
Flask==0.10.1
Flask-Login==0.3.2
Flask-Mandrill==0.2
Flask-Migrate==1.6.0
Flask-MySQL==1.3
Flask-Script==2.0.5
Flask-SQLAlchemy==2.0
Flask-WTF==0.12
itsdangerous==0.24
Jinja2==2.8
Mako==1.0.2
MarkupSafe==0.23
MySQL-python==1.2.5
pbr==1.8.1
python-editor==0.4
requests==2.8.1
six==1.10.0
SQLAlchemy==1.0.8
sqlalchemy-migrate==0.10.0
sqlparse==0.1.16
Tempita==0.5.2
Werkzeug==0.10.4
wheel==0.24.0
<<<<<<< HEAD
<<<<<<< HEAD
WTForms==2.0.2
=======
>>>>>>> master
=======
WTForms==2.0.2
sqlalchemy-migrate==0.10.0
>>>>>>> master
